fix: Patch dropdown bug
